Log out before accepting an invitation
If a user is already logged in, then accepting an invitation by clicking
on the invitation link would lead to weird behaviour. Instead of
displaying the sign-up page as expected, the user may be redirected to
another page. The invitation may still be accepted though, leaving to an
undesired state where the invitation is accepted, but the new user
account has not been created.

For these reasons, it is best to log out any existing user sesion before
accepting an invitation and displaying the sign-up page.