Add invitee field to AbstractBaseInvitation model
This library has been designed to keep Invitation instances in the
database, even after the invitation has been accepted and the user
account has been created. The only reason to keep these invitations
around is to have a track record of invitations. It seems likely in that
case that the inviter would want to be able to interact with the new
user account, depending on the application. They may want to visit their
profile page or something.

Before this commit, this was impossible to do, as there was nothing
linking the invitation instance to a user instance. At first glance, it
might seem that an email address would link both models, but that is not
the case. A user is not obligated to sign up with the same email address
that was used to invite them. Even if they do use the same email
address, they may change their email address in their user account later
on. So an email address could not be relied on to link these two models.

This commit modifies the modify to add a new field: "invitee". It sets
this field after signup. This requires that allauth be installed.

Fixes #252